fix(bundle): keep the lens catalogue out of client bundles
CompareCollections is a client component and imported @/lib/collections, which
called getAllLenses for its membership counts. That pulled the catalogue JSON,
its zod schema, and data.ts into the compare page's client bundle: a 391KB chunk
holding all 217 lenses with every field, for code that never read it.

Splits collections into the half that takes its lenses as arguments and
collection-stats, which needs the catalogue. CollectionPills now receives
slug -> count as a prop (43 numbers, ~683 bytes) instead of counting members
itself, so the one client path into the catalogue is gone.

server-only on data.ts, lens-link-index.ts and collection-stats.ts turns this
class of leak into a build error. It caught one the split alone had missed:
CollectionPills has no "use client" of its own but is rendered by
CompareCollections, so it was in the client graph too.

compare's first-load JS: 1876KB -> 1250KB.
