+++ package.json
@@ -45,6 +45,7 @@
     "react-markdown": "^10.1.0",
     "remark-cjk-friendly": "^2.3.1",
     "remark-gfm": "^4.0.1",
+    "server-only": "^0.0.1",
     "shadcn": "^4.1.0",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.5.0",
