+++ src/app/[locale]/lenses/[mount]/(browse)/collections/[slug]/page.tsx
@@ -3,7 +3,8 @@ import { notFound } from "next/navigation";
 import { getTranslations, setRequestLocale } from "next-intl/server";
 import { ArrowRight, Flag } from "lucide-react";
 import { Link } from "@/i18n/navigation";
-import { COLLECTIONS, getCollectionStats, getRelatedCollectionsWithStats } from "@/lib/collections";
+import { COLLECTIONS } from "@/lib/collections";
+import { getCollectionStats, getRelatedCollectionsWithStats } from "@/lib/collection-stats";
 import { getLensesByMount } from "@/lib/lens/data";
 import { urlSegmentToMount, mountSeoLabel, mountHasCollections } from "@/lib/mount";
 import { buildAlternates, defaultOgImages } from "@/lib/seo";
